Extracts functionality to modules Removes unnecessary dependencies
